[nexus] Add `DataStore::instance_fetch_all`
In order to implement the `instance-update` saga ongoing in #5749, it's
necessary to have a way to query the database for the state of an
instance, its active VMM, its target VMM, and active migration in a
single, atomic query. This will be used as a snapshot of the instance
state that the update saga will operate on.

This commit adds an implementation of such a query in the form of a
`DataStore::instance_fetch_all` method, returning an `InstanceSnapshot`
struct that bundles together all these records. We do this by `LEFT
JOIN`ing the `instance` table with the `vmm` and `migration` table on
the `instance`'s various ID columns, with a query that should look
something like:

```sql
SELECT * FROM instance
LEFT JOIN vmm
    ON ((instance.active_propolis_id = vmm.id)
    OR (instance.target_propolis_id = vmm.id))
LEFT JOIN migration ON (instance.migration_id = migration.id)
WHERE instance.id = instance_id
```

Ideally, we would instead do two subqueries that join `instance` with
`vmm` on the active and destination IDs separately and return a single
row which contains both in a way that can be distinguished in the result
set, but I wasn't able to figure out how to do that in Diesel. Instead,
the query may return two rows, if the instance has both an active and
target VMM, and iterate over the result set to find both VMMs in Rust
code. This is a bit uglier, but it works fine.