Randomize read-only parent UUIDs during migration
During an instance migration, the migration destination is given a new
volume checkout for a disk during the migration saga. Checkout increases
the generation number in the sub volume portion of the Volume but not
the read-only parent portion. This means that the migration destination
can activate and take over the read/write ownership of the sub volumes,
as it should.

However, the read-only side of the hierarchy does not have a generation
number bump. This is ok, it doesn't need one: there's no exclusive
access that needs to be maintained for the read-only data, and many
different Upstairs can connect to a read-only downstairs without issues.

But for read-only connections that share an upstairs ID, the read-only
Crucible Downstairs will kick out which ever connected first, and
subsequent connections will kick out previous ones. This is a problem in
the migration case, because the read-only parents will be the same for
the source and the destination instance's disks - volume IDs are _not_
randomized during checkout.

This is only problem during migration - for other cases, the read-only
parent's UUIDs are properly randomized. This commits adds a
randomization step during `instance_ensure_registered` for the migration
case only: the source and destination read-only Upstairs should be
uniquely identified.

Also, this commit truncates the long strings that were preventing `cargo
fmt` from formatting instance.rs, and bundles those changes in.