refactor: use typed instance and Propolis IDs
As part of the ongoing Instance Lifecycle Reimagination Exercise of
2024, we'd like to refactor sled agent's instance manager so that sled
agent thinks purely in terms of the VMMs it's managing and not in terms
of the instances that those VMMs incarnate. To make this less
challenging, add a Propolis UUID type, then rework a bunch of code to
distinguish properly between instance IDs and Propolis IDs, both in sled
agent and in Nexus.

In general the change stops at the database model boundary; that is, I
didn't try to make any of the DB model types use typed UUIDs (though I'm
curious to see if one could). Beyond that there are admittedly few rules
to describe what changed--I started with a couple of types in sled agent
and let the refactoring katamari roll from there.