Add `migration` table and explicit migration tracking in sled-agent (#5859)
As part of ongoing work on improving instance lifecycle management (see
#5749) , we intend to remove the `InstanceRuntimeState` tracking from
`sled-agent`, and make Nexus the sole owner of `instance` records in
CRDB, with a new `instance-update` saga taking over the responsibility
of managing the instance's state transitions.

In order to properly manage the instance state machine, Nexus will need
information about the status of active migrations that are currently
only available to sled-agents. For example, if an instance is migrating,
and a sled agent reports that the source VMM is `Destroyed`, Nexus
doesn't presently have the capability to determine whether the source
VMM was destroyed because the migration completed successfully, or that
the source shut down prior to starting the migration, resulting in a
failure.

In order for Nexus to correctly manage state updates during live
migration, we introduce a new `migration` table to the schema for
tracking the state of ongoing migrations. The `instance-migrate` saga
creates a `migration` record when beginning a migration. The Nexus and
sled-agent APIs are extended to include migration state updates from
sled-agents to Nexus. In *this* branch, the `migration` table is
(basically) write-only; Nexus doesn't really read from it, and just
stuffs updates into it. In the future, however, this will be used by the
`instance-update` saga.

It occurred to me that, in addition to using the migration table for
instance updates, it might also be useful to add an OMDB command to look
up the status of a migration using this table. However, I decided that
made more sense as a follow-up change, as I'd like to get back to
integrating this into #5749.

Fixes #2948