Only apply external IP allowlist to Nexus (#5900)
- Fixes #5892
- Modifies the application of the external services IP allowlist so that
it's only relevant for Nexus API servers, rather than all
external-facing services (DNS being the other example today). It is not
always possible to know the peer addresses for DNS servers in the case
of recursive DNS, and so the allowlist cannot directly apply to external
DNS. This works by inserting the allowlist entries as a host-filter,
which we were doing before, but only on the named VPC Firewall rule for
the Nexus VPC Subnet.