Account for region snapshots during allocation
When replacing a snapshot (aka a read-only downstairs somewhere), region
allocation must occur for a snapshot volume. A snapshot volume's region
set is currently only composed of read-only downstairs which all contain
the same data, and any newly allocated region should take those into
account when looking to meet the redundancy criteria for a region set:
in production, allocate the new region to a distinct sled.

This is done by slightly changing the region allocation query to accept
an optional snapshot id, and, if that is supplied, will add the region
snapshot's pools to the `existing_zpools` temporary table in order to
prevent region allocation there.