Fix "project" (silo) IP pool view perms for non-admins (#5887)
Closes #5883

Authz for IP pools is undercooked. Every user has `CreateChild` on them
so they can allocate IPs, but they generally don't have `Read` on them
unless they're a fleet viewer. Ideally, we'd be able to say "you have
`Read` on an IP pool if it is linked to your silo", but I don't know how
to express that relationship with polar (plus it requires a join to the
silo-pool links table to tell).

Update: after
[discussing](https://matrix.to/#/!YNYPOVxjAUeXksTcRj:oxide.computer/$XTbZiYkEFRKWHwe6XSVivBrqiWWl_z25u2_TcjwRohE?via=oxide.computer&via=unix.house&via=matrix.org)
this with @davepacheco in chat, we agreed on a direction but also agreed
it might be fairly complicated. Based on that, I think it would be most
expedient to

1. Merge this fix as-is (with a link to
https://github.com/oxidecomputer/omicron/issues/3995, which I forgot I
had made 10 months ago 😢)
2. Update #3995 with more details from the chat