BUGFIX: use sled filter to apply v2p mappings
It has been reported that marking a sled as "non_provisionable"
causes it to lose the instance v2p mappings. This is a regression.
It was determined that the scoping of the mappings is too strict,
relying on the sled_policy to be set to "in service" via the db view.

The preferred option is to rely on the sled policy machinery to do the
filtering for us, so the view has been removed and the querying logic
updated accordingly.

Also, based on a comment in a previous related PR, we are now selecting
target sleds for instances using the active_propolis_id to avoid
potential conflicts during live migrations.