Add PEM encoded certificate to external api response (#5078)
* Add pem encoded certificate to certificate list and view endpoints

When managing multiple silos, it quickly becomes desirable to be able to
use the oxide cli and/or sdk to manage them. Handling certificate
rotation currently requires storing external state about what
certificate is registered with which silo as our external API only
returns identity metadata about the certificate. This PR proposes adding
the the public certificate to the fields returned for a certificate.

This initial implementation tries to take the simplest approach to this:
1. Return the exact representation that we store in the database
(PEM-encoded string)
2. Returns the certificate on both list and view endpoints

An alternative approach may be to only return the full certificate as a
part of the view endpoint, but I'm not sure if that is a pattern that we
use.

The other consideration is how this should handle the failure to
construct a String from the bytes form of a certificate. Currently it is
very disruptive, and hard to debug if bad data were to be stored in the
database.

---------

Co-authored-by: reflector[bot] <123+reflector[bot]@users.noreply.github.com>