be more consistent about workspace members (#5921)
1. We have reasons for excluding some packages from the default
workspace members, but there was nothing checking that newly-added
packages were added to both. I added an additional check to `cargo xtask
check-workspace-deps` to ensure that `default-members` is equal to
`members` minus xtask and end-to-end-tests. This had the effect that the
tests in `api_identity` were not being run in CI!
2. Add `--workspace` to `cargo clippy` and `cargo doc` in CI to run
checks across the entire tree.
3. Fix clippy lints / doc errors that were not being caught in the
previous setup.