Increase default session timeout to 8 hours idle, 24 hours absolute (#5920)
We want to make the values configurable (#5477), but in the meantime we
can improve the user experience substantially by increasing the
hard-coded values so people get logged out less often. I picked 8
hours/24 hours so that someone could log in in the morning and stay
logged in all day.

It's worth noting that there are no deleterious effects if the IdP's own
session timeout is less than ours as long as the customer is willing to
accept the longer console session. It just means that whenever the user
gets logged out of the console and wants to log back in, they may have
to also complete the login step in their IdP.