Define timeseries schema in TOML (#5889)
This is the first commit in a thread of work to support updates to
timeseries schema. In our first step, we're moving the definition of
timeseries from the Rust structs we use today, to TOML text files. Each
file describes one target and all the metrics associated with it, from
which we generate exactly the same Rust code that it replaces.

This opens the door to a lot of downstream work, including well-defined
updates to schema; detection of conflicting schema; improved metadata
for timeseries; and generation of documentation. As an example, this
moves exactly one (1) timeseries into the new format, the physical
datalink statistics tracked through illumos kstats.

- Move the `oximeter` crate into a private implementation crate, and
re-export it at its original path
- Add intermediate representation of timeseries schema, and code for
deserializing TOML timeseries definitions, checking them for conflicts,
and handling updates. Note that **no actual updates** are currently
supported. The ingesting code includes a check that version numbers are
exactly 1, except in tests. We include tests that check updates in a
number of ways, but until the remainder of the work is done, we limit
timeseries in the wild to their current version of 1.
- Add a macro for consuming timeseries definitions in TOML, and
generating the equivalent Rust code. Developers should use this new
`oximeter::use_timeseries!()` proc macro to create new timeseries.
- Add an integration test in Nexus that pulls in timeseries schema
ingested with with `oximeter::use_timeseries!()`, and checks them all
for conflicts. As developers add new timeseries in TOML format, they
must be added here as well to ensure we detect problems at CI time.
- Updates `kstat-rs` dep to simplify platform-specific code. This is
part of this commit because we need the definitions of the
physical-data-link timeseries for our integration test, but those were
previously only compiled on illumos platforms. They're now included
everywhere, while the tests remain illumos-only.