[#3886 4/4] Drive and finish a region replacement (#5885)
When a disk is expunged, any region that was on that disk is assumed to
be gone. A single disk expungement can put many Volumes into degraded
states, as one of the three mirrors of a region set is now gone. Volumes
that are degraded in this way remain degraded until a new region is
swapped in, and an Upstairs performs the necessary repair operation
(either through a Live Repair or Reconciliation). Nexus can only
initiate these repairs - it does not participate in them, instead
requesting that an Upstairs perform the repair.

These repair operations can only be done by an Upstairs running as part
of an activated Volume: either Nexus has to send this Volume to a Pantry
and repair it there, or Nexus has to talk to a propolis that has that
active Volume. Further complicating things is that the Volumes in
question can be activated and deactivated as a result of user action,
namely starting and stopping Instances. This will interrupt any on-going
repair. This is ok! Both operations support being interrupted, but as a
result it's then Nexus' job to continually monitor these repair
operations and initiate further operations if the current one is
interrupted.

A single saga invocation is not enough to continually make sure a Volume
is being repaired, so driving one of the repair operations forward
happens as a saga that is triggered from a background task: this is
called the _region replacement drive_ saga. It wll:

- transition a region replacement request to state Driving, again
blocking out other invocations of the same saga

- check if Nexus has taken an action to initiate a repair yet. if not,
then one is needed. if it has previously initiated a repair operation,
the state of the system is examined: is that operation still running?
has something changed? further action may be required depending on this
observation.

- if an action is required, Nexus will prepare an action that will
initiate either Live Repair or Reconciliation based on the current
observed state of the system.

- that action is then executed. if there was an error, then the saga
unwinds. if it was successful, it is recorded as a "repair step" in CRDB
and will be checked the next time the saga runs.

- if Nexus observed an Upstairs telling it that a repair was completed
or not necessary, then the request is placed into the ReplacementDone
state, otherwise it is placed back into the Running state. if the saga
unwinds, it unwinds back to the Running state.

The background task responsible for triggering the drive saga will also
scan for notifications of a successful live repair or reconciliation,
and transition region replacement requests to ReplacementDone if it sees
one.

If a region replacement request is in state ReplacementDone, _region
replacement finish_ saga is triggered, which will:

- transition a request into Completing

- delete the old region by deleting a transient Volume that refers to it

- transition the request to the Complete state

For the entire region replacement work, testing was done manually using
the Canada region using the following test cases:

- a disk needing repair is attached to a instance for the duration of
the repair

- a disk needing repair is attached to a instance that is migrated
mid-repair

- a disk needing repair is attached to a instance that is stopped
mid-repair

- a disk needing repair is attached to a instance that is stopped
mid-repair, then started in the middle of the pantry's repair

- a detached disk needs repair

- a detached disk needs repair, and is then attached to an instance that
is then started

- a sled is expunged, causing region replacement requests for all
regions on it

Fixes #3886
Fixes #5191