VPC Subnet Routing [2/2] -- Custom Routers and NIC 'transit IP' lists (#5823)
This PR builds on #5777 to provide the Custom routers for subnets as
described in RFD21. This entails a few things:
* We remove the `unpublished = true` tag from the user API for VPC
routers and routes.
* Custom routers may be attached/detached to a VPC subnet using the
`custom_router` field in subnet `POST` and `PUT` requests.
* NICs now individually have a `transit_ips` list, which denotes an
additional set of CIDR blocks that a NIC is allowed to send and receive
traffic on. This is set during `POST` and/or `PUT` on instances which
are stopped. This is a key feature to enable software routing by
instances, as today's default behaviour drops any packets not matching
an assigned IP for an instance.
* I suspect there will be some discussion over the shape of this API, so
there isn't yet test coverage here until we know we're happy with it.
* Revisited which router routes can be created by users, e.g., better
validation on v4/v6 dest/target pairs.

There are some allowances around currently non-existent features:
* **Internet Gateways.** We allow unlimited use of one pseudo-gateway,
`inetgw:outbound`, which appears in our existing rules. Using this
target sends packets upstream as it does today.
* **VPC peering.** VPCs as destinations/targets are currently disallowed
in router routes.

Closes #2116.