review feedback from @gjcolombo
This changes how the instance update saga is constructed to better
handle cases where the active VMM was destroyed *because* a migration
out was successful, which should be treated as a "migration success"
and should not release virtual provisioning resources or delete the
instance's oximeter producer. Also, the instance record is now written
back in the same query as releasing the lock, to avoid potential race
conditions.