nicer `instance_set_migration_ids`
This commit adds a subquery for ensuring that the active VMM is in the
`Running` or `Rebooting` states, so that we know it's okay to migrate
out of when we start the migration.

Additionally, this changes the query to return the instance record, so
that the saga action `sim_set_migration_ids` can avoid a refetch query.