[nexus] switch nexus internal API to a trait, add scaffolding to manage it (#5844)
Part of the work on [RFD 479 Dropshot API traits](https://rfd.shared.oxide.computer/rfd/0479).

As a first, impactful example, this PR switches over the Nexus internal
API to being defined by a trait. Some minor refactoring was required but
overall it took me around half an hour to do this.

Included also is a tool, `cargo xtask openapi`, which is used to manage
this and future documents, as well as instructions that folks can
hopefully follow.