reconfigurator: Ensure all durable datasets, not just crucible (#6065)
Reconfigurator currently ensures a `Dataset` row exists for all crucible
zones, but fails to do so for other zones with durable datasets (notably
Cockroach, which we now support adding and expunging!). This PR fixes
that.

I think we also need reconfigurator to learn how to _delete_ dataset
rows for expunged zones, right? That will be a followup PR.