[installinator] convert artifact/progress API to a trait (#6074)
To retain consistency with all of the other APIs, I've changed the name
to just `installinator-api`. `artifact` was the case at first but the same API
is now also used for progress reporting, so it's no longer quite accurate.

We had this indirection going on that's no longer necessary. We also don't need
to have the installinator-artifactd binary that generated the OpenAPI document,
since the manager takes care of all of that.