[nix flake] make dev shell moldy
This commit switches the dev shell provided by the Nix flake to use
`mold` as the linker, rather than `ld`. This results in substantially
faster build times (at least on my build machine, where building Nexus
with `ld` results in memory bottlenecks and swaps to disk a bunch).

To use `mold` in the Nix dev shell, we override the Nix `stdenv` to use
Mold as the system linker, as well as overriding `clang` as the C
compiler (which we did previously).

@sunshowers I'd be interested to know whether the Nix dev shell is now
usable on your system, as I know we had problems previously due to your
system-wide non-Nix use of `mold`...