[wicketd] convert wicket -> wicketd API to a trait (#6078)
Move the wicketd API to a new `wicketd-api` crate, and use that
everywhere.

I had to do a few more changes along the way:

1. In the interest of keeping the API crate itself small, I moved a lot of the
underlying types to wicket-common. An example is the update-engine spec
for preflight checks -- that shouldn't really live in the API crate.
2. As a followup I also used those types in the API rather than relying
on generic specs. This made the JSON a bit bigger but resulted in a nice
cleanup on the wicket end, since the types can be shared.
3. With wicketd-api in its own crate, two copies of `SpIdentifier` and all of the
other types around that got way out of hand. Luckily, addressing that
was quite simple -- just using the gateway-client copy of `SpIdentifier`
everywhere. This means a few more `replace`s, but I think the value of
never having to see an `SpIdentifier2` is quite high.

The wicketd API being difficult to update was a key motivation behind me
starting work on API traits, so it's nice to see that come to fruition.