move update saga needed logic into saga module
I could be convinced otherwise about this change. My thought process was
that it seemed nice for that logic to live next to the documentation for
the update saga itself, so that a reader can see it all at the same
time. But, on the other hand, this is a function that's only called in
one place, which is also not my favorite thing.