[nexus-db-queries] make saga_update more resilient, record_event idempotent (#6113)
See discussion in https://github.com/oxidecomputer/omicron/issues/2416
and https://github.com/oxidecomputer/omicron/issues/6090#issuecomment-2229509411.

A summary of the changes here:

1. Made `saga_create_event` idempotent. Previously, creating another
event that duplicated the one which already existed would fail -- now it
succeeds. These events are meant to be an append-only idempotent log, so
this is okay. Also added a test for this.
2. `saga_update_state` was already idempotent -- added a test which made
sure of this. Also added a comment about how idempotence may not be
enough in the future.
3. Added a retry loop around saga state updates, similar to the one
around recording events.