Add sled-identifiers to instance vCPU stats (#6128)
This adds the sled UUID and baseboard info into the existing vCPU stats.
This is an incompatible change, and will require dropping old data from
the previous versions of the timeseries. That's unused today, and it's
not yet clear how we'll actually manage the timeseries data across a
schema change, so I'd like to include this now.