[sled-agent] Self assembling switch zone (#5593)
## Overview

This PR migrates the switch zone to a self assembling format. There are
a few bits of old code I'll be cleaning up, more logs I'll be adding and
some documentation about how the switch zone flow works, but I'll do
this in follow up PRs to keep this one as compact as possible.

## Caveats

I've tested this in a local single node deployment and in the a4x2
testbed. Unfortunately, this is not enough testing to make sure all of
the services play nice together on a real rack. We'll have to keep an
eye on dogfood when this is deployed.

The only services that depend on the [common
networking](https://github.com/oxidecomputer/omicron/blob/30eb1ee38987201ac71f2115fdd89da4b08710c7/zone-setup/src/bin/zone-setup.rs#L578-L690)
[service](https://github.com/oxidecomputer/omicron/blob/30eb1ee38987201ac71f2115fdd89da4b08710c7/smf/zone-network-setup/manifest.xml)
are dendrite and MGS. While this makes sense on the a4x2 testbed, I'd
like to verify that these dependencies make sense when running on a real
rack.

As several people have worked on different parts of this zone, I've
tagged a whole bunch of people for review, sorry if this is overkill!
Just want to make sure I've got the right eyes on each service of the
zone.

Related: https://github.com/oxidecomputer/omicron/issues/1898
Closes: https://github.com/oxidecomputer/omicron/issues/2884

TODO:

- [x] Update Dendrite hashes after merging
https://github.com/oxidecomputer/dendrite/pull/990