Add a omdb db volume command, with VCR support
Added a subcommand to dump the volume field in the database.
The volume's data field holds a string that should be a
VolumeConstructionRequest.  This also adds the ability to dump out
a subset of the fields in that VCR.

Updated the `db disks info` and `db snapshots info` subcommands
to print out the VCR as well.