[#5333 1/6] Snapshot replacement models + queries
At a high level, snapshot replacement looks similar to region
replacement, but there are distinct differences. The biggest difference
is that replacing a read-only Downstairs does _not_ start any operation
in the corresponding Upstairs: the data is the same before and after, so
no live repair or reconciliation is required.

Snapshot replacement operates at the level of a region snapshot (which
1:1 corresponds to a read-only Downstairs running out of a .zfs/snapshot
directory), and is intended to replace this with a read-only region.
It starts by allocating a replacement region using the new `source`
parameter to invoke a clone operation (which will copy all the blocks
from another read-only Downstairs in the region set), and once that's
done replace the snapshot's address with the new region's in the
snapshot volume.

A snapshot's volume construction request is never directly used to
construct a volume - a copy is made, and used as the read-only parent
for something else like a disk or an image. After the target replacement
in the previous paragraph occurs, any copy made will pick up the region
set that includes the new read-only region. However, any previous copy
will be referencing the snapshot's address. The next phase of snapshot
replacement finds all the volumes that reference the address of the
snapshot being replaced, and create a "snapshot replacement step" record
for those volumes: Nexus needs to perform the same snapshot address ->
new region address replacement in each Upstairs that was constructed
where the old snapshot address appears somewhere in the Volume
hierarchy, and this is tracked in these "step" records.

A snapshot replacement is considered finished when there's no more
references to the snapshot being replaced, and all the "step" records
are in their terminal state.

Start by separating out the DB models, queries, and schema changes
required:

1. snapshot replacement records

Similar to region replacement records, snapshot replacement records
transition through a set of states during the whole snapshot replacement
process:

         Requested   <--
                       |
             |         |
             v         |
                       |
         Allocating  --

             |
             v

       ReplacementDone  <--
                          |
             |            |
             v            |
                          |
     DeletingOldVolume  --

             |
             v

          Running

             |
             v

         Complete

which are captured in the `SnapshotReplacementState` enum.

First, transitioning from Requested through to ReplacementDone is the
responsibility of a new "snapshot replacement start" saga. This saga is
triggered by a new background task (creatively titled "snapshot
replacement start").

Second, transitioning from ReplacementDone through to Running is the
responsibility of the "snapshot replacement garbage collect" saga. This
is also triggered by a second new background task (creatively titled
"snapshot replacement garbage collect").

Last, transitioning from Running to Complete is done in a third new
background task titled "snapshot replacement finish". Note that this is
_not_ done by a saga.

State transitions are performed in a similar way to how region
replacement record states are transitioned.

2. snapshot replacement step records

Once a snapshot's volume has had the relevant target replaced, any still
running Upstairs that was constructed with that target requires a
replacement call to be sent. A "step" record is created for each of
these volumes, and will transition through the following states:

        Requested   <--
                      |
            |         |
            v         |
                      |
         Running    --

            |
            v

        Complete

            |
            v

      VolumeDeleted

Transitioning from Requested through to Complete is the responsibility
of the "snapshot replacement step" saga, and transitioning from
Complete to VolumeDeleted is the responsibility of the "snapshot
replacement step garbage collect" saga.

Only when all step records for a request are in state "VolumeDeleted"
will a snapshot replacement record transition to "Complete".

3. also, the `volume_replace_snapshot" function

The `volume_replace_snapshot` function is also included in this PR.